chore(html): add treelist templates
