+++ packages/html/src/treelist/index.ts
@@ -1,1 +1,6 @@
 export * from './treelist.spec';
+export * from './templates/treelist-normal';
+export * from './templates/treelist-with-column-menu';
+export * from './templates/treelist-with-filter-menu';
+export * from './templates/treelist-row-reordering';
+export * from './templates/treelist-aggregates';
