+++ packages/html/src/treelist/templates/treelist-with-column-menu.tsx
@@ -1,0 +1,6 @@
+import { TreeListNormal } from "./treelist-normal";
+
+export const TreeListWithColumnMenu = (props) => (
+    <TreeListNormal menu="column" {...props} />
+);
+
