+++ packages/html/src/treelist/templates/treelist-with-filter-menu.tsx
@@ -1,0 +1,6 @@
+import { TreeListNormal } from "./treelist-normal";
+
+export const TreeListWithFilterMenu = (props) => (
+    <TreeListNormal menu="filter" {...props} />
+);
+
