+++ packages/html/src/treelist/tests/treelist.tsx
@@ -1,32 +1,33 @@
 import { Icon } from '../../icon';
-import { TreeList } from '../../treelist';
-import { GridHeader, GridHeaderTable, GridHeaderCell, GridContainer, GridContent, GridTable } from '../../grid';
+import { GridHeader, GridHeaderTable, GridHeaderCell, GridContent, GridTable } from '../../grid';
 import { TableThead, TableTbody, TableRow, TableTd } from '../../table';
+import { TreeListNormal, TreeListWithColumnMenu, TreeListWithFilterMenu } from '..';
 
 export default () =>(
     <>
         <div id="test-area" className="k-d-grid">
             <section>
-
-                <TreeList id="treelist" className="k-grid-no-scrollbar">
-                    <GridHeader style={{ paddingRight: '0px' }}>
-                        <div className="k-grid-header-wrap">
-                            <GridHeaderTable>
-                                <colgroup>
-                                    <col style={{ width: "250px" }}/>
-                                    <col />
-                                </colgroup>
-                                <TableThead>
-                                    <TableRow>
-                                        <GridHeaderCell menu="column" columnTitle="First Name" className="k-filterable"></GridHeaderCell>
-                                        <GridHeaderCell menu="column" columnTitle="Last Name" className="k-filterable"></GridHeaderCell>
-                                    </TableRow>
-                                </TableThead>
-                            </GridHeaderTable>
-                        </div>
-                    </GridHeader>
-                    <GridContainer>
-                        <GridContent className="k-auto-scrollable">
+                TreeList without Scrollbar
+                <TreeListNormal className="k-grid-no-scrollbar"
+                    gridHeader={
+                        <GridHeader style={{ paddingRight: '0px' }}>
+                            <div className="k-grid-header-wrap">
+                                <GridHeaderTable>
+                                    <colgroup>
+                                        <col style={{ width: "250px" }}/>
+                                        <col />
+                                    </colgroup>
+                                    <TableThead>
+                                        <TableRow>
+                                            <GridHeaderCell menu="column" filterable columnTitle="First Name"></GridHeaderCell>
+                                            <GridHeaderCell menu="column" filterable columnTitle="Last Name"></GridHeaderCell>
+                                        </TableRow>
+                                    </TableThead>
+                                </GridHeaderTable>
+                            </div>
+                        </GridHeader>
+                    } gridContent={
+                        <GridContent>
                             <GridTable>
                                 <colgroup>
                                     <col style={{ width: "250px" }}/>
@@ -43,97 +44,109 @@ export default () =>(
                                 </TableTbody>
                             </GridTable>
                         </GridContent>
-                    </GridContainer>
-                </TreeList>
-
+                    }
+                />
             </section>
 
             <section>
-                <TreeList id="treelist">
+                TreeList Normal
+                <TreeListNormal></TreeListNormal>
+            </section>
+
+            <section>
+                TreeList with Column Menu
+                <TreeListWithColumnMenu gridContent={
+                    <GridContent>
+                        <GridTable>
+                            <colgroup>
+                                <col />
+                                <col />
+                                <col />
+                            </colgroup>
+                            <TableTbody>
+                                <TableRow>
+                                    <TableTd>
+                                        <Icon className="k-treelist-toggle" icon="caret-alt-right"></Icon>
+                                        Daryl
+                                    </TableTd>
+                                    <TableTd>Sweeney</TableTd>
+                                    <TableTd>CEO</TableTd>
+                                </TableRow>
+                            </TableTbody>
+                        </GridTable>
+                    </GridContent>
+                }/>
+            </section>
+
+            <section>
+                TreeList with Filter Menu
+                <TreeListWithFilterMenu gridContent={
+                    <GridContent>
+                        <GridTable>
+                            <colgroup>
+                                <col />
+                                <col />
+                                <col />
+                            </colgroup>
+                            <TableTbody>
+                                <TableRow>
+                                    <TableTd>
+                                        <Icon className="k-treelist-toggle" icon="caret-alt-right"></Icon>
+                                        Daryl
+                                    </TableTd>
+                                    <TableTd>Sweeney</TableTd>
+                                    <TableTd>CEO</TableTd>
+                                </TableRow>
+                            </TableTbody>
+                        </GridTable>
+                    </GridContent>
+                }/>
+            </section>
+
+            <section>
+                TreeList Sorting
+                <TreeListNormal gridHeader={
                     <GridHeader>
                         <div className="k-grid-header-wrap">
                             <GridHeaderTable>
                                 <colgroup>
-                                    <col style={{ width: "280px" }} />
-                                    <col style={{ width: "160px" }} />
+                                    <col />
+                                    <col />
                                     <col />
                                 </colgroup>
                                 <TableThead>
                                     <TableRow>
-                                        <GridHeaderCell menu="filter" columnTitle="First Name"></GridHeaderCell>
-                                        <GridHeaderCell menu="filter" columnTitle="Last Name"></GridHeaderCell>
-                                        <GridHeaderCell menu="filter" columnTitle="Position"></GridHeaderCell>
+                                        <GridHeaderCell columnTitle="Name"></GridHeaderCell>
+                                        <GridHeaderCell columnTitle="Position" sortable></GridHeaderCell>
+                                        <GridHeaderCell columnTitle="Hire Date"></GridHeaderCell>
                                     </TableRow>
                                 </TableThead>
                             </GridHeaderTable>
                         </div>
                     </GridHeader>
-                    <GridContainer>
-                        <GridContent className="k-auto-scrollable">
-                            <GridTable>
-                                <colgroup>
-                                    <col style={{ width: "280px" }} />
-                                    <col style={{ width: "160px" }} />
-                                    <col />
-                                </colgroup>
-                                <TableTbody>
-                                    <TableRow className="k-treelist-group">
-                                        <TableTd>
-                                            <Icon className="k-treelist-toggle" icon="caret-alt-down"></Icon>
-                                            Daryl
-                                        </TableTd>
-                                        <TableTd>Sweeney</TableTd>
-                                        <TableTd>CEO</TableTd>
-                                    </TableRow>
-                                    <TableRow className="k-treelist-group k-alt" alt>
-                                        <TableTd>
-                                            <Icon className="k-treelist-toggle" icon="none"></Icon>
-                                            <Icon className="k-treelist-toggle" icon="caret-alt-down"></Icon>
-                                            Guy
-                                        </TableTd>
-                                        <TableTd>Wooten</TableTd>
-                                        <TableTd>Chief Technical Officer</TableTd>
-                                    </TableRow>
-                                    <TableRow className="k-treelist-group">
-                                        <TableTd>
-                                            <Icon className="k-treelist-toggle" icon="none"></Icon>
-                                            <Icon className="k-treelist-toggle" icon="none"></Icon>
-                                            <Icon className="k-treelist-toggle" icon="caret-alt-down"></Icon>
-                                            Buffy
-                                        </TableTd>
-                                        <TableTd>Weber</TableTd>
-                                        <TableTd>VP, Engineering</TableTd>
-                                    </TableRow>
-                                    <TableRow className="k-treelist-group k-alt" alt>
-                                        <TableTd>
-                                            <Icon className="k-treelist-toggle" icon="none"></Icon>
-                                            <Icon className="k-treelist-toggle" icon="none"></Icon>
-                                            <Icon className="k-treelist-toggle" icon="none"></Icon>
-                                            <Icon className="k-treelist-toggle" icon="caret-alt-down"></Icon>
-                                            Hyacinth
-                                        </TableTd>
-                                        <TableTd>Hood</TableTd>
-                                        <TableTd>Team Lead</TableTd>
-                                    </TableRow>
-                                    <TableRow>
-                                        <TableTd>
-                                            <Icon className="k-treelist-toggle" icon="none"></Icon>
-                                            <Icon className="k-treelist-toggle" icon="none"></Icon>
-                                            <Icon className="k-treelist-toggle" icon="none"></Icon>
-                                            <Icon className="k-treelist-toggle" icon="none"></Icon>
-                                            <Icon className="k-treelist-toggle" icon="none"></Icon>
-                                            Akeem
-                                        </TableTd>
-                                        <TableTd>Carr</TableTd>
-                                        <TableTd>Junior Software Developer</TableTd>
-                                    </TableRow>
-                                </TableTbody>
-                            </GridTable>
-                        </GridContent>
-                    </GridContainer>
-                </TreeList>
+                }
+                gridContent={
+                    <GridContent>
+                        <GridTable>
+                            <colgroup>
+                                <col />
+                                <col className="k-sorted"/>
+                                <col />
+                            </colgroup>
+                            <TableTbody>
+                                <TableRow>
+                                    <TableTd>
+                                        <Icon className="k-treelist-toggle" icon="caret-alt-right"></Icon>
+                                        Daryl
+                                    </TableTd>
+                                    <TableTd>Sweeney</TableTd>
+                                    <TableTd>CEO</TableTd>
+                                </TableRow>
+                            </TableTbody>
+                        </GridTable>
+                    </GridContent>
+                }/>
             </section>
         </div>
-
     </>
 );
